refactor: deduplicate resource concurrency control
